Refactored CanvasBase
I noticed that the logic with an empty extra layer
at the end of the Canvas is only required to create
the iterator, without knowing the end item in the list.

So I refactored the IndependentContext iterator and
removed the requirement for an empty layer at the end
of the Canvas.

- Extracted CanvasBase parent as a private variable
- CanvasBase type changed from list to vector (this allowed
to find additional problems)
- IndependentContext refactored to support current behaviour